fix(perf): drop redundant pid() casts + wrap over-long LOST-flush line
Dev's `Process::pid()` returns `u32`, so the `as u32` casts in the per-task
arm path and overflow handler are redundant (clippy unnecessary_cast). Also
wrap the PERF_RECORD_LOST flush store to satisfy the newer rustfmt width.

+++ os/StarryOS/kernel/src/perf/sampling.rs
@@ -446,7 +446,7 @@ pub fn pmu_overflow_handler(_ctx: IrqContext) -> IrqReturn {
                 None => {
                     let curr = ax_task::current();
                     match curr.try_as_thread() {
-                        Some(thr) => (thr.proc_data.proc.pid() as u32, thr.tid()),
+                        Some(thr) => (thr.proc_data.proc.pid(), thr.tid()),
                         None => {
                             let id = curr.id().as_u64() as u32;
                             (id, id)
@@ -498,7 +498,8 @@ pub fn pmu_overflow_handler(_ctx: IrqContext) -> IrqReturn {
                     if unsafe { ring_write(ring_vaddr, ring_len, &lost_rec) } {
                         // SAFETY: as above.
                         unsafe {
-                            (*(lost_reported_ptr as *const AtomicU64)).store(total, Ordering::Relaxed)
+                            (*(lost_reported_ptr as *const AtomicU64))
+                                .store(total, Ordering::Relaxed)
                         };
                     }
                 }
